requirements.txt
opencv-python
Pillow
requests
replicate
moondream==0.0.6
moviepy
